requirements.txt
numpy 
pandas 
scikit-learn 
google-cloud-storage
matplotlib
seaborn